+++ mod-source-record-storage-server/src/main/java/org/folio/rest/impl/InitAPIImpl.java
@@ -15,10 +15,10 @@ import org.apache.logging.log4j.Logger;
 import org.folio.config.ApplicationConfig;
 import org.folio.processing.events.EventManager;
 import org.folio.rest.resource.interfaces.InitAPI;
+import org.folio.services.handlers.HoldingsPostProcessingEventHandler;
 import org.folio.services.handlers.InstancePostProcessingEventHandler;
 import org.folio.services.handlers.MarcAuthorityEventHandler;
 import org.folio.services.handlers.MarcBibliographicMatchEventHandler;
-import org.folio.services.handlers.MarcHoldingsEventHandler;
 import org.folio.services.handlers.actions.ModifyRecordEventHandler;
 import org.folio.spring.SpringContextUtil;
 import org.folio.verticle.consumers.DataImportConsumersVerticle;
@@ -36,6 +36,9 @@ public class InitAPIImpl implements InitAPI {
   private InstancePostProcessingEventHandler instancePostProcessingEventHandler;
 
   @Autowired
+  private HoldingsPostProcessingEventHandler holdingsPostProcessingEventHandler;
+
+  @Autowired
   private ModifyRecordEventHandler modifyRecordEventHandler;
 
   @Autowired
@@ -43,9 +46,6 @@ public class InitAPIImpl implements InitAPI {
 
   @Autowired
   private MarcAuthorityEventHandler marcAuthorityEventHandler;
-
-  @Autowired
-  private MarcHoldingsEventHandler marcHoldingsEventHandler;
 
   @Value("${srs.kafka.ParsedMarcChunkConsumer.instancesNumber:1}")
   private int parsedMarcChunkConsumerInstancesNumber;
@@ -77,10 +77,10 @@ public class InitAPIImpl implements InitAPI {
 
   private void registerEventHandlers() {
     EventManager.registerEventHandler(instancePostProcessingEventHandler);
+    EventManager.registerEventHandler(holdingsPostProcessingEventHandler);
     EventManager.registerEventHandler(modifyRecordEventHandler);
     EventManager.registerEventHandler(marcBibliographicMatchEventHandler);
     EventManager.registerEventHandler(marcAuthorityEventHandler);
-    EventManager.registerEventHandler(marcHoldingsEventHandler);
   }
 
   private Future<?> deployConsumerVerticles(Vertx vertx) {